+++ index.html
@@ -5,7 +5,6 @@
     <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>No Sleep Drive</title>
-
   </head>
   <body>
     <div id="root"></div>
